Plugin framework refactoring
This commit adds a new PluginManager that controls loading of all plugins,
regardless of type (kernel vs. stage, static vs. shared). Subsequently, the
StageFactory and KernelFactory classes are greatly simplified as there was
significant redundant code. Macros are provided to simplify creation of each
type of plugin, but are not explicitly required. Each plugin registers it's
name, description, link, type, and will soon include a version.

Dropped program_options in pdal.cpp and added ability to selectively load
plugin types (or none at all).

+++ plugins/hexbin/test/HexbinFilterTest.cpp
@@ -71,11 +71,11 @@ TEST(HexbinFilterTest, HexbinFilterTest_test_1)
         "use in situations where you do not want to estimate based on "
         "a sample");
 
-    ReaderPtr reader(f.createReader("readers.las"));
+    std::unique_ptr<Stage> reader(f.createStage("readers.las"));
     EXPECT_TRUE(reader.get());
     reader->setOptions(options);
 
-    FilterPtr hexbin(f.createFilter("filters.hexbin"));
+    std::unique_ptr<Stage> hexbin(f.createStage("filters.hexbin"));
     EXPECT_TRUE(hexbin.get());
     hexbin->setOptions(options);
     hexbin->setInput(reader.get());
@@ -93,8 +93,11 @@ TEST(HexbinFilterTest, HexbinFilterTest_test_1)
     printChildren(out, m);
     out.close();
 
-    EXPECT_TRUE(Support::compare_text_files(filename,
-        Support::datapath("filters/hexbin.txt")));
+    bool ok = Support::compare_text_files(filename,
+        Support::datapath("filters/hexbin.txt"));
 
-    FileUtils::deleteFile(filename);
+    if (ok)
+        FileUtils::deleteFile(filename);
+
+    EXPECT_TRUE(ok);
 }
